+++ test/waveform_generator_test.py
@@ -20,8 +20,9 @@ class TestWaveformGeneratorInstantiationWithoutOptionalParameters(unittest.TestC
 
     def setUp(self):
         self.waveform_generator = \
-            tupak.gw.waveform_generator.WaveformGenerator(1, 4096,
-                                                          frequency_domain_source_model=dummy_func_dict_return_value)
+            tupak.gw.waveform_generator.WaveformGenerator(
+                1, 4096,
+                frequency_domain_source_model=dummy_func_dict_return_value)
         self.simulation_parameters = dict(amplitude=1e-21, mu=100, sigma=1,
                                           ra=1.375,
                                           dec=-1.2108,
@@ -34,16 +35,14 @@ class TestWaveformGeneratorInstantiationWithoutOptionalParameters(unittest.TestC
 
     def test_repr(self):
         expected = 'WaveformGenerator(duration={}, sampling_frequency={}, start_time={}, ' \
-                   'frequency_domain_source_model={}, time_domain_source_model={}, parameters={}, ' \
-                   'parameter_conversion={}, non_standard_sampling_parameter_keys={}, waveform_arguments={})'\
+                   'frequency_domain_source_model={}, time_domain_source_model={}, ' \
+                   'parameter_conversion={}, waveform_arguments={})'\
             .format(self.waveform_generator.duration,
                     self.waveform_generator.sampling_frequency,
                     self.waveform_generator.start_time,
                     self.waveform_generator.frequency_domain_source_model.__name__,
                     self.waveform_generator.time_domain_source_model,
-                    self.waveform_generator.parameters,
                     None,
-                    self.waveform_generator.non_standard_sampling_parameter_keys,
                     self.waveform_generator.waveform_arguments)
         self.assertEqual(expected, repr(self.waveform_generator))
 
@@ -52,16 +51,14 @@ class TestWaveformGeneratorInstantiationWithoutOptionalParameters(unittest.TestC
             tupak.gw.waveform_generator.WaveformGenerator(1, 4096,
                                                           time_domain_source_model=dummy_func_dict_return_value)
         expected = 'WaveformGenerator(duration={}, sampling_frequency={}, start_time={}, ' \
-                   'frequency_domain_source_model={}, time_domain_source_model={}, parameters={}, ' \
-                   'parameter_conversion={}, non_standard_sampling_parameter_keys={}, waveform_arguments={})'\
+                   'frequency_domain_source_model={}, time_domain_source_model={}, ' \
+                   'parameter_conversion={}, waveform_arguments={})'\
             .format(self.waveform_generator.duration,
                     self.waveform_generator.sampling_frequency,
                     self.waveform_generator.start_time,
                     self.waveform_generator.frequency_domain_source_model,
                     self.waveform_generator.time_domain_source_model.__name__,
-                    self.waveform_generator.parameters,
                     None,
-                    self.waveform_generator.non_standard_sampling_parameter_keys,
                     self.waveform_generator.waveform_arguments)
         self.assertEqual(expected, repr(self.waveform_generator))
 
@@ -71,16 +68,14 @@ class TestWaveformGeneratorInstantiationWithoutOptionalParameters(unittest.TestC
 
         self.waveform_generator.parameter_conversion = conversion_func
         expected = 'WaveformGenerator(duration={}, sampling_frequency={}, start_time={}, ' \
-                   'frequency_domain_source_model={}, time_domain_source_model={}, parameters={}, ' \
-                   'parameter_conversion={}, non_standard_sampling_parameter_keys={}, waveform_arguments={})'\
+                   'frequency_domain_source_model={}, time_domain_source_model={}, ' \
+                   'parameter_conversion={}, waveform_arguments={})'\
             .format(self.waveform_generator.duration,
                     self.waveform_generator.sampling_frequency,
                     self.waveform_generator.start_time,
                     self.waveform_generator.frequency_domain_source_model.__name__,
                     self.waveform_generator.time_domain_source_model,
-                    self.waveform_generator.parameters,
                     conversion_func.__name__,
-                    self.waveform_generator.non_standard_sampling_parameter_keys,
                     self.waveform_generator.waveform_arguments)
         self.assertEqual(expected, repr(self.waveform_generator))
 
@@ -100,6 +95,7 @@ class TestWaveformGeneratorInstantiationWithoutOptionalParameters(unittest.TestC
         self.assertIsInstance(self.waveform_generator.time_array, np.ndarray)
 
     def test_source_model_parameters(self):
+        self.waveform_generator.parameters = self.simulation_parameters.copy()
         self.assertListEqual(sorted(list(self.waveform_generator.parameters.keys())),
                              sorted(list(self.simulation_parameters.keys())))
 
@@ -136,14 +132,15 @@ class TestSetters(unittest.TestCase):
         del self.simulation_parameters
 
     def test_parameter_setter_sets_expected_values_with_expected_keys(self):
-        self.waveform_generator.parameters = self.simulation_parameters
+        self.waveform_generator.parameters = self.simulation_parameters.copy()
         for key in self.simulation_parameters:
             self.assertEqual(self.waveform_generator.parameters[key], self.simulation_parameters[key])
 
     def test_parameter_setter_none_handling(self):
-        self.waveform_generator.parameters = None
-        self.assertListEqual(sorted(list(self.waveform_generator.parameters.keys())),
-                             sorted(list(self.simulation_parameters.keys())))
+        with self.assertRaises(TypeError):
+            self.waveform_generator.parameters = None
+        # self.assertListEqual(sorted(list(self.waveform_generator.parameters.keys())),
+        #                      sorted(list(self.simulation_parameters.keys())))
 
     def test_frequency_array_setter(self):
         new_frequency_array = np.arange(1, 100)
@@ -157,11 +154,13 @@ class TestSetters(unittest.TestCase):
 
     def test_parameters_set_from_frequency_domain_source_model(self):
         self.waveform_generator.frequency_domain_source_model = dummy_func_dict_return_value
+        self.waveform_generator.parameters = self.simulation_parameters.copy()
         self.assertListEqual(sorted(list(self.waveform_generator.parameters.keys())),
                              sorted(list(self.simulation_parameters.keys())))
 
     def test_parameters_set_from_time_domain_source_model(self):
         self.waveform_generator.time_domain_source_model = dummy_func_dict_return_value
+        self.waveform_generator.parameters = self.simulation_parameters.copy()
         self.assertListEqual(sorted(list(self.waveform_generator.parameters.keys())),
                              sorted(list(self.simulation_parameters.keys())))
 
@@ -195,10 +194,10 @@ class TestFrequencyDomainStrainMethod(unittest.TestCase):
     def test_parameter_conversion_is_called(self):
         self.waveform_generator.parameter_conversion = MagicMock(side_effect=KeyError('test'))
         with self.assertRaises(KeyError):
-            self.waveform_generator.frequency_domain_strain()
+            self.waveform_generator.frequency_domain_strain(
+                parameters=self.simulation_parameters)
 
     def test_frequency_domain_source_model_call(self):
-        self.waveform_generator.parameters = self.simulation_parameters
         expected = self.waveform_generator.frequency_domain_source_model(self.waveform_generator.frequency_array,
                                                                          self.simulation_parameters['amplitude'],
                                                                          self.simulation_parameters['mu'],
@@ -207,36 +206,39 @@ class TestFrequencyDomainStrainMethod(unittest.TestCase):
                                                                          self.simulation_parameters['dec'],
                                                                          self.simulation_parameters['geocent_time'],
                                                                          self.simulation_parameters['psi'])
-        actual = self.waveform_generator.frequency_domain_strain()
+        actual = self.waveform_generator.frequency_domain_strain(
+            parameters=self.simulation_parameters)
         self.assertTrue(np.array_equal(expected['plus'], actual['plus']))
         self.assertTrue(np.array_equal(expected['cross'], actual['cross']))
 
     def test_time_domain_source_model_call_with_ndarray(self):
         self.waveform_generator.frequency_domain_source_model = None
         self.waveform_generator.time_domain_source_model = dummy_func_array_return_value
-        self.waveform_generator.parameters = self.simulation_parameters
 
         def side_effect(value, value2):
             return value
 
         with mock.patch('tupak.core.utils.nfft') as m:
             m.side_effect = side_effect
-            expected = self.waveform_generator.time_domain_strain()
-            actual = self.waveform_generator.frequency_domain_strain()
+            expected = self.waveform_generator.time_domain_strain(
+                parameters=self.simulation_parameters)
+            actual = self.waveform_generator.frequency_domain_strain(
+                parameters=self.simulation_parameters)
             self.assertTrue(np.array_equal(expected, actual))
 
     def test_time_domain_source_model_call_with_dict(self):
         self.waveform_generator.frequency_domain_source_model = None
         self.waveform_generator.time_domain_source_model = dummy_func_dict_return_value
-        self.waveform_generator.parameters = self.simulation_parameters
 
         def side_effect(value, value2):
             return value, self.waveform_generator.frequency_array
 
         with mock.patch('tupak.core.utils.nfft') as m:
             m.side_effect = side_effect
-            expected = self.waveform_generator.time_domain_strain()
-            actual = self.waveform_generator.frequency_domain_strain()
+            expected = self.waveform_generator.time_domain_strain(
+                parameters=self.simulation_parameters)
+            actual = self.waveform_generator.frequency_domain_strain(
+                parameters=self.simulation_parameters)
             self.assertTrue(np.array_equal(expected['plus'], actual['plus']))
             self.assertTrue(np.array_equal(expected['cross'], actual['cross']))
 
@@ -244,7 +246,8 @@ class TestFrequencyDomainStrainMethod(unittest.TestCase):
         self.waveform_generator.time_domain_source_model = None
         self.waveform_generator.frequency_domain_source_model = None
         with self.assertRaises(RuntimeError):
-            self.waveform_generator.frequency_domain_strain()
+            self.waveform_generator.frequency_domain_strain(
+                parameters=self.simulation_parameters)
 
     def test_key_popping(self):
         self.waveform_generator.parameter_conversion = MagicMock(return_value=(dict(amplitude=1e-21, mu=100, sigma=1,
@@ -253,7 +256,8 @@ class TestFrequencyDomainStrainMethod(unittest.TestCase):
                                                                                     psi=2.659, c=None, d=None),
                                                                                ['c', 'd']))
         try:
-            self.waveform_generator.frequency_domain_strain()
+            self.waveform_generator.frequency_domain_strain(
+                parameters=self.simulation_parameters)
         except RuntimeError:
             pass
         self.assertListEqual(sorted(self.waveform_generator.parameters.keys()),
@@ -279,10 +283,10 @@ class TestTimeDomainStrainMethod(unittest.TestCase):
     def test_parameter_conversion_is_called(self):
         self.waveform_generator.parameter_conversion = MagicMock(side_effect=KeyError('test'))
         with self.assertRaises(KeyError):
-            self.waveform_generator.time_domain_strain()
+            self.waveform_generator.time_domain_strain(
+                parameters=self.simulation_parameters)
 
     def test_time_domain_source_model_call(self):
-        self.waveform_generator.parameters = self.simulation_parameters
         expected = self.waveform_generator.time_domain_source_model(self.waveform_generator.time_array,
                                                                     self.simulation_parameters['amplitude'],
                                                                     self.simulation_parameters['mu'],
@@ -291,36 +295,39 @@ class TestTimeDomainStrainMethod(unittest.TestCase):
                                                                     self.simulation_parameters['dec'],
                                                                     self.simulation_parameters['geocent_time'],
                                                                     self.simulation_parameters['psi'])
-        actual = self.waveform_generator.time_domain_strain()
+        actual = self.waveform_generator.time_domain_strain(
+            parameters=self.simulation_parameters)
         self.assertTrue(np.array_equal(expected['plus'], actual['plus']))
         self.assertTrue(np.array_equal(expected['cross'], actual['cross']))
 
     def test_frequency_domain_source_model_call_with_ndarray(self):
         self.waveform_generator.time_domain_source_model = None
         self.waveform_generator.frequency_domain_source_model = dummy_func_array_return_value
-        self.waveform_generator.parameters = self.simulation_parameters
 
         def side_effect(value, value2):
             return value
 
         with mock.patch('tupak.core.utils.infft') as m:
             m.side_effect = side_effect
-            expected = self.waveform_generator.frequency_domain_strain()
-            actual = self.waveform_generator.time_domain_strain()
+            expected = self.waveform_generator.frequency_domain_strain(
+                parameters=self.simulation_parameters)
+            actual = self.waveform_generator.time_domain_strain(
+                parameters=self.simulation_parameters)
             self.assertTrue(np.array_equal(expected, actual))
 
     def test_frequency_domain_source_model_call_with_dict(self):
         self.waveform_generator.time_domain_source_model = None
         self.waveform_generator.frequency_domain_source_model = dummy_func_dict_return_value
-        self.waveform_generator.parameters = self.simulation_parameters
 
         def side_effect(value, value2):
             return value
 
         with mock.patch('tupak.core.utils.infft') as m:
             m.side_effect = side_effect
-            expected = self.waveform_generator.frequency_domain_strain()
-            actual = self.waveform_generator.time_domain_strain()
+            expected = self.waveform_generator.frequency_domain_strain(
+                parameters=self.simulation_parameters)
+            actual = self.waveform_generator.time_domain_strain(
+                parameters=self.simulation_parameters)
             self.assertTrue(np.array_equal(expected['plus'], actual['plus']))
             self.assertTrue(np.array_equal(expected['cross'], actual['cross']))
 
@@ -328,7 +335,8 @@ class TestTimeDomainStrainMethod(unittest.TestCase):
         self.waveform_generator.time_domain_source_model = None
         self.waveform_generator.frequency_domain_source_model = None
         with self.assertRaises(RuntimeError):
-            self.waveform_generator.time_domain_strain()
+            self.waveform_generator.time_domain_strain(
+                parameters=self.simulation_parameters)
 
     def test_key_popping(self):
         self.waveform_generator.parameter_conversion = MagicMock(return_value=(dict(amplitude=1e-2,
@@ -339,7 +347,8 @@ class TestTimeDomainStrainMethod(unittest.TestCase):
                                                                                     psi=2.659, c=None, d=None),
                                                                                ['c', 'd']))
         try:
-            self.waveform_generator.time_domain_strain()
+            self.waveform_generator.time_domain_strain(
+                parameters=self.simulation_parameters)
         except RuntimeError:
             pass
         self.assertListEqual(sorted(self.waveform_generator.parameters.keys()),